[update] mainView의 하위 View 설정
- profileView
- infoView
- tdView
