--- AirPlan/AirPlan/MainInfoView.swift
@@ -1,17 +1,0 @@
-//
-//  MainInfoView.swift
-//  AirPlan
-//
-//  Created by 이중엽 on 2023/06/12.
-//
-
-import UIKit
-import SnapKit
-
-class MainInfoView: UIView, UIEssentials {
-    func initialSetup() {
-    }
-    
-    func makeUI() {
-    }
-}
